Fix for orientation of handle cross section, and manifold issue.
Orientation simply required a 90 degree rotation to be inserted.
Manifold issue is trickier - had to interpolate towards radial coordinates
at the ends of the handles, so they definitely intersected with the mesh.
Also had to remove the end caps - they are not planar and it was confusing
Blender, so it works better without them. This does however make it
really sensitive to the handles not intersecting, as if that happens you
will lose chunks of the output mesh that are not sealed, which could be
everything.